requirements.txt
attrs==25.4.0
certifi==2025.10.5
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
py==1.11.0
pyparsing==3.2.5
pytest==8.4.2
toml==0.10.2
Pillow
